Ruchira's changes for inner contracts

+++ src/postgresrep/postgreshandler.go
@@ -11,7 +11,6 @@ import "database/sql"
 
 func InitialMigrationC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket, couchViewName, xmlPath, enableDelete string) {
 
-	isCommit:= false
 	file, _ := os.Create("loginsert.txt")
 	//getting table mappings
 	var tables = GetXMLData(xmlPath, file)
@@ -93,19 +92,12 @@ func InitialMigrationC2PG(dbname, user, password, host, couchHost, couchPool, co
 			if (err != nil) || (f == nil) || (reflect.ValueOf(f).Kind() != reflect.Map) {
 				fmt.Println("Object not found or mismacthed with structures for key :" + res.Rows[i].ID)
 			} else {
-
 				fmt.Println("Got value for key :" + res.Rows[i].ID + " and type is " + reflect.TypeOf(f).Kind().String())
 
 				for _, table := range tables.Tables {
 					//file.WriteString("Iterating table table.CouchName " + "\n")
 					fmt.Println("Iterating table " + table.CouchName + "\n")
-					if(table.SkipType == "true"){
-						isCommit = true
-					}else{
-						isCommit = strings.Contains(res.Rows[i].ID, table.CouchName)
-					}
-					
-					if isCommit == true {
+					if strings.Contains(res.Rows[i].ID, table.CouchName) {
 						m := f.(map[string]interface{})
 						//fmt.Println(m)
 						var insertQuery = table.PGInsert
@@ -122,38 +114,67 @@ func InitialMigrationC2PG(dbname, user, password, host, couchHost, couchPool, co
 								}
 							}
 						}
-
+						
+						var strArray = []string{}
 						fmt.Println("Iterating nested columns " + "\n")
+						for _, nested := range table.NestedColumn {
+							strArray = strings.Split(nested.ColumnName, ".")
+							break
+						}
+						
+						var nestedTrue = true
+						
+						for i := 0; i < reflect.ValueOf(m[strArray[0]]).Len(); i++ {
+							fmt.Println("Index Number")
+							fmt.Println(i)
+							// reset the insert query for nested columns. without reset query, data not replace with query parameters
+							var insertQuery = table.PGInsert
+						
 						for _, nested := range table.NestedColumn {
 							var strArray = []string{}
 							//file.WriteString("Iterating nested column " + nested.ColumnName +"\n")
 							strArray = strings.Split(nested.ColumnName, ".")
 							var nestedValue interface{}
 							if len(strArray) == 3 {
-								index, _ := strconv.Atoi(strArray[1])
+								//index, _ := strconv.Atoi(strArray[1])
 								if reflect.ValueOf(m[strArray[0]]).Len() > 0 {
-									nestedValue = m[strArray[0]].([]interface{})[index].(map[string]interface{})[strArray[2]]
-									if nested.Fixed == 1 {
-										if nestedValue != nil {
-											if reflect.TypeOf(nestedValue).Kind() == reflect.Float64 {
-												accountInt := int64(nestedValue.(float64))
-												stringVal := strconv.FormatInt(accountInt, 10)
-												//file.WriteString(nested.ColumnName+" has fixed")
-												insertQuery = strings.Replace(insertQuery, "@"+nested.ColumnName, stringVal, -1)
+									//--------------------------------------------------------------------------------------------------------------------------------							
+									
+									
+										//nestedValue = m[strArray[0]].([]interface{})[index].(map[string]interface{})[strArray[2]]
+										nestedValue = m[strArray[0]].([]interface{})[i].(map[string]interface{})[strArray[2]]
+										fmt.Println(nestedValue)
+										
+										
+										if nested.Fixed == 1 {
+											if nestedValue != nil {
+												if reflect.TypeOf(nestedValue).Kind() == reflect.Float64 {
+													accountInt := int64(nestedValue.(float64))
+													stringVal := strconv.FormatInt(accountInt, 10)
+													//file.WriteString(nested.ColumnName+" has fixed")
+													insertQuery = strings.Replace(insertQuery, "@"+nested.ColumnName, stringVal, -1)
+												}
+											} else {
+												file.WriteString(nested.ColumnName + " value is null")
 											}
 										} else {
-											file.WriteString(nested.ColumnName + " value is null")
+											//file.WriteString(nested.ColumnName+" has not fixed")
+											fmt.Println("Not Fixed " +nested.ColumnName)
+											fmt.Println(insertQuery)
+											fmt.Println("-----------------------------")
+											insertQuery = strings.Replace(insertQuery, "@"+nested.ColumnName, GetStringValue(nestedValue), -1)
+											fmt.Println(insertQuery)
+											fmt.Println("Insert should changed")
 										}
-									} else {
-										//file.WriteString(nested.ColumnName+" has not fixed")
-										insertQuery = strings.Replace(insertQuery, "@"+nested.ColumnName, GetStringValue(nestedValue), -1)
-									}
+									
+									
+									//------------------------------------------------------------------------------------------------------------------------------------
 								} else {
 									insertQuery = strings.Replace(insertQuery, "@"+nested.ColumnName, "", -1)
 									//file.WriteString(insertQuery)
 								}
 
-							} else if len(strArray) == 2 {
+							} else if len(strArray) == 2 { fmt.Println("array 2 condition")
 								index, _ := strconv.Atoi(strArray[1])
 								if reflect.ValueOf(m[strArray[0]]).Len() < 1 {
 									insertQuery = strings.Replace(insertQuery, "@"+nested.ColumnName, "", -1)
@@ -175,30 +196,66 @@ func InitialMigrationC2PG(dbname, user, password, host, couchHost, couchPool, co
 							}
 
 						}
+						
+						//insert query here
+						
+						nestedTrue = false
+						fmt.Println(nestedTrue)
+						//------------------------------------------------
+							fmt.Println(insertQuery)
+							result, err := db.Exec(insertQuery)
+							file.WriteString(insertQuery)
+						
+							if err != nil {
+								fmt.Println("Postgres insertion error : " + err.Error())
+								file.WriteString(err.Error() + "\n")
+							} else {
+								status, ok := result.RowsAffected()
+								if ok == nil {
+									fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+									file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+								} else {
+									fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+									file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+								}
+							}
+						
+						//------------------------------------------------
+						
+						} //index loop.....
 
 						for k, v := range m {
-
+							fmt.Println("------------------ range loop --------------------")
+							fmt.Println(insertQuery)
+							fmt.Println(k)
 							if v != nil {
+								
+								fmt.Println(v)
 								insertQuery = strings.Replace(insertQuery, "@"+k, GetStringValue(v), -1)
 							} else {
 								insertQuery = strings.Replace(insertQuery, "@"+k, "", -1)
 							}
 						}
-
-						fmt.Println(insertQuery)
-						result, err := db.Exec(insertQuery)
-						file.WriteString(insertQuery)
-						if err != nil {
-							fmt.Println("Postgres insertion error : " + err.Error())
-							file.WriteString(err.Error() + "\n")
-						} else {
-							status, ok := result.RowsAffected()
-							if ok == nil {
-								fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
-								file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+						
+						if nestedTrue { 
+							fmt.Println(nestedTrue)
+							fmt.Println("nestedQuery can't come here")
+							fmt.Println(insertQuery)
+							result, err := db.Exec(insertQuery)
+							file.WriteString(insertQuery)
+						
+							if err != nil {
+								fmt.Println("Postgres insertion error : " + err.Error())
+								file.WriteString(err.Error() + "\n")
 							} else {
-								fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
-								file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+								status, ok := result.RowsAffected()
+								if ok == nil {
+									fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+									file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+								} else {
+									fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+									file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+								}
 							}
 						}
 
@@ -214,9 +271,10 @@ func InitialMigrationC2PG(dbname, user, password, host, couchHost, couchPool, co
 		file.WriteString(strconv.Itoa(len(res.Rows)) + " data processed" + "\n")
 		skipCount += len(res.Rows)
 		file.Close()
+		bucket.Close()
+		db.Close()
 	}
 
-	bucket.Close()
 	db.Close()
 }
 
@@ -288,12 +346,12 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 	for skipCount < totalCouchRows {
 
 		file.WriteString("Getting couch data skip " + strconv.Itoa(skipCount) + "\n")
-		if enableDelete == "false" {
+		if(enableDelete == "false"){
 			skipDeleteCount = skipCount
-		} else {
+		}else{
 			skipDeleteCount = 0
 		}
-
+		
 		res, err := bucket.View(couchViewName, couchViewName, map[string]interface{}{
 			"stale": false,
 			"limit": 5000,
@@ -309,8 +367,8 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 
 		for i := 0; i < len(res.Rows); i++ {
 
-			updateId := reflect.ValueOf(res.Rows[i].Value).Interface().(map[string]interface{})["ID"].(string)
-			updateType := reflect.ValueOf(res.Rows[i].Value).Interface().(map[string]interface{})["Type"].(string)
+			updateId := reflect.ValueOf(res.Rows[0].Value).Interface().(map[string]interface{})["ID"].(string)
+			updateType := reflect.ValueOf(res.Rows[0].Value).Interface().(map[string]interface{})["Type"].(string)
 			bucketGet.Get(updateId, &f)
 
 			if updateType == "Insert" {
@@ -338,25 +396,33 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 								//file.WriteString("Iterating change coumn " + prop.ColumnName + "\n")
 								var propValue = m[prop.ColumnName]
 								var accountInt int64
-								if propValue != nil {
-									if reflect.TypeOf(propValue).Kind() == reflect.Float64 {
-										accountInt = int64(m[prop.ColumnName].(float64))
-										stringVal := strconv.FormatInt(accountInt, 10)
-										insertQuery = strings.Replace(insertQuery, "@"+prop.ColumnName+"_", stringVal, 100)
-									}
-								} else {
-									insertQuery = strings.Replace(insertQuery, "@"+prop.ColumnName+"_", "", 100)
+								if reflect.TypeOf(propValue).Kind() == reflect.Float64 {
+									accountInt = int64(m[prop.ColumnName].(float64))
+									stringVal := strconv.FormatInt(accountInt, 10)
+									insertQuery = strings.Replace(insertQuery, "@"+prop.ColumnName+"_", stringVal, 100)
 								}
 
 							}
+							
+							
+							var nestedTrue = true
+							
+							var strArray = []string{}
+							fmt.Println("Iterating nested columns " + "\n")
+							for _, nested := range table.NestedColumn {
+								strArray = strings.Split(nested.ColumnName, ".")
+								break
+							}
+							
+							for i := 0; i < reflect.ValueOf(m[strArray[0]]).Len(); i++ {
 
 							for _, nested := range table.NestedColumn {
 								var strArray = []string{}
 								strArray = strings.Split(nested.ColumnName, ".")
 								if len(strArray) == 3 {
-									index, _ := strconv.Atoi(strArray[1])
+									//index, _ := strconv.Atoi(strArray[1])
 									if reflect.ValueOf(m[strArray[0]]).Len() > 0 {
-										nestedValue := m[strArray[0]].([]interface{})[index].(map[string]interface{})[strArray[2]]
+										nestedValue := m[strArray[0]].([]interface{})[i].(map[string]interface{})[strArray[2]]
 
 										if nested.Fixed == 1 {
 											if reflect.TypeOf(nestedValue).Kind() == reflect.Float64 {
@@ -389,7 +455,33 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 
 								}
 
-							}
+							  }
+							
+							
+								//insert query here
+						
+								nestedTrue = false
+								fmt.Println(nestedTrue)
+								//------------------------------------------------
+								fmt.Println(insertQuery)
+								result, err := db.Exec(insertQuery)
+								file.WriteString(insertQuery)
+						
+								if err != nil {
+									fmt.Println("Postgres insertion error : " + err.Error())
+									file.WriteString(err.Error() + "\n")
+								} else {
+									status, ok := result.RowsAffected()
+									if ok == nil {
+										fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+										file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+									} else {
+										fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+										file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+									}
+								}
+							
+							} // index loop endup here
 
 							for k, v := range m {
 
@@ -400,32 +492,39 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 								}
 							}
 
-							result, err := db.Exec(insertQuery)
-							fmt.Println(insertQuery)
-							if err != nil {
-								fmt.Println("Postgres insertion error : " + err.Error())
-								file.WriteString(err.Error() + "\n")
-							} else {
-								status, ok := result.RowsAffected()
-								if ok == nil {
-									fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
-									file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+							if nestedTrue{ 
+								fmt.Println(nestedTrue)
+								fmt.Println("nestedQuery can't come here")
+							
+								result, err := db.Exec(insertQuery)
+								if enableDelete == "true" {
+									bucket.Delete(res.Rows[i].ID)
+									fmt.Println(res.Rows[i].ID + " has deleted from bucket " + couchBucket)
+								}
+								fmt.Println(insertQuery)
+								if err != nil {
+									fmt.Println("Postgres insertion error : " + err.Error())
+									file.WriteString(err.Error() + "\n")
 								} else {
-									fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
-									file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+									status, ok := result.RowsAffected()
+									if ok == nil {
+										fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+										file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + strconv.FormatInt(status, 10))
+									} else {
+										fmt.Println("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+										file.WriteString("Migrate status of key " + res.Rows[0].ID + " is " + ok.Error())
+									}
 								}
 							}
 
 							fmt.Println("processed " + strconv.Itoa(i) + " out of " + strconv.Itoa(res.TotalRows))
-							break
 						} else {
 							fmt.Println("Skipped key " + res.Rows[i].ID + "processed " + strconv.Itoa(i) + " out of " + strconv.Itoa(res.TotalRows))
+							if enableDelete == "true" {
+								bucket.Delete(res.Rows[i].ID)
+								fmt.Println(res.Rows[i].ID + " has deleted from bucket " + couchBucket)
+							}
 						}
-					}
-
-					if enableDelete == "true" {
-						bucket.Delete(res.Rows[i].ID)
-						fmt.Println(res.Rows[i].ID + " has deleted from bucket " + couchBucket)
 					}
 				}
 
@@ -476,7 +575,7 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 									updateQuery = strings.Replace(updateQuery, "@"+nested.ColumnName, "", -1)
 								}
 
-							} else if len(strArray) == 2 {
+							} else if len(strArray) == 2 { fmt.Println("array 2 condition")
 								index, _ := strconv.Atoi(strArray[1])
 								if reflect.ValueOf(m[strArray[0]]).Len() > 0 {
 									nestedValue := m[strArray[0]].([]interface{})[index]
@@ -497,8 +596,9 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 						}
 
 						for k, v := range m {
-
+							fmt.Println("------------------ range loop ---------------------------")
 							if v != nil {
+								fmt.Println(k)
 								updateQuery = strings.Replace(updateQuery, "@"+k, GetStringValue(v), -1)
 							} else {
 								updateQuery = strings.Replace(updateQuery, "@"+k, "", -1)
@@ -507,6 +607,10 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 
 						//file.WriteString(updateQuery)
 						result, err := db.Exec(updateQuery)
+						if enableDelete == "true" {
+							bucket.Delete(res.Rows[i].ID)
+							fmt.Println(res.Rows[i].ID + " has deleted from bucket " + couchBucket)
+						}
 						fmt.Println(updateQuery)
 						if err != nil {
 							fmt.Println("Postgres update error : " + err.Error())
@@ -520,17 +624,17 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 								fmt.Println("Migrate status of key " + updateId + " is " + ok.Error())
 								file.WriteString("Migrate status of key " + updateId + " is " + ok.Error())
 							}
-							break;
 						}
 					} else {
 						fmt.Println("Skipped key " + updateId + "processed " + strconv.Itoa(i) + " out of " + strconv.Itoa(res.TotalRows))
+						if enableDelete == "true" {
+							bucket.Delete(res.Rows[i].ID)
+							fmt.Println(res.Rows[i].ID + " has deleted from bucket " + couchBucket)
+						}
 					}
+
 				}
 
-				if enableDelete == "true" {
-					bucket.Delete(res.Rows[i].ID)
-					fmt.Println(res.Rows[i].ID + " has deleted from bucket " + couchBucket)
-				}
 			}
 		}
 
@@ -540,7 +644,6 @@ func UpdateC2PG(dbname, user, password, host, couchHost, couchPool, couchBucket,
 	}
 
 	file.Close()
-	//closing log file
 	fmt.Println("Log file closed")
 	db.Close()
 	fmt.Println("Postgres connection closed")
